+++ supabase/functions/_cmd/cmd.ts
@@ -1,8 +1,7 @@
-import { generateKey } from '../_shared/utils/encrypt.ts';
 import { log } from '../_shared/utils/log.ts';
 import { runScheduler } from './scheduler/scheduler.ts';
 import { runWorker } from './worker/worker.ts';
-import { generateAppleSecretKey } from './tools/tools.ts';
+import { runTools } from './tools/tools.ts';
 
 /**
  * Next to the Supabase Edge functions we also have to create an command which
@@ -26,47 +25,12 @@ const main = (args: string[]) => {
       log('error', 'Worker crashed', { error: err.toString() });
       Deno.exit(1);
     });
-  } else if (
-    args.length === 2 && args[0] === 'tools' &&
-    args[1] === 'generate-key'
-  ) {
-    /**
-     * The "tools generate-key" command can be invoked via the following
-     * command:
-     *   deno run --allow-net --allow-env --import-map=./supabase/functions/import_map.json ./supabase/functions/_cmd/cmd.ts tools generate-key
-     */
-    generateKey().then((data) => {
-      log('info', 'Encryption key was generated', {
-        key: data.rawKey,
-        iv: data.iv,
-      });
+  } else if (args.length >= 2 && args[0] === 'tools') {
+    log('info', 'Start tools...');
+    runTools(args).then(() => {
       Deno.exit(0);
     }).catch((err) => {
-      log('error', 'Failed to generate encryption key', {
-        error: err.toString(),
-      });
-      Deno.exit(1);
-    });
-  } else if (
-    args.length === 6 && args[0] === 'tools' &&
-    args[1] === 'generate-apple-secret-key'
-  ) {
-    /**
-     * The "tools generate-key" command can be invoked via the following
-     * command:
-     *   deno run --allow-env --allow-read --import-map=./supabase/functions/import_map.json ./supabase/functions/_cmd/cmd.ts tools generate-apple-secret-key <KEY-ID> <TEAM-ID> <SERVICE-ID> <FILE>
-     */
-    generateAppleSecretKey(args[2], args[3], args[4], args[5]).then((data) => {
-      log('info', 'Encryption key was generated', {
-        kid: data.kid,
-        exp: new Date(data.exp * 1000).toString(),
-        jwt: data.jwt,
-      });
-      Deno.exit(0);
-    }).catch((err) => {
-      log('error', 'Failed to generate encryption key', {
-        error: err.toString(),
-      });
+      log('error', 'Tools crashed', { error: err.toString() });
       Deno.exit(1);
     });
   } else {
